feat: refactor ps module to have more clarity
